+++ src/messagingSummaryItem/MessagingSummaryItem.story.mdx
@@ -1,0 +1,25 @@
+import { Meta, Story, Preview, Props } from '@storybook/addon-docs/blocks'
+
+import { MessagingSummaryItem } from './index'
+
+<Meta title="MessagingSummaryItem" />
+
+# **MessagingSummaryItem**
+
+<Preview>
+  <Story name="Default">
+    <MessagingSummaryItem />
+  </Story>
+</Preview>
+
+## Specifications
+
+## Usage
+
+```jsx
+import { MessagingSummaryItem } from '@blablacar/ui-library/build/MessagingSummaryItem'
+
+<MessagingSummaryItem />
+```
+
+<Props of={MessagingSummaryItem} />
