+++ src/messagingSummaryItem/MessagingSummaryItem.style.tsx
@@ -1,0 +1,3 @@
+import styled from 'styled-components'
+
+export const StyledMessagingSummaryItem = styled.div``
